feat: Split Benchmark API functions and corresponding tests into distinct files
