+++ platform/src/apis/Platform.Api.Benchmark/Constants.cs
@@ -6,6 +6,10 @@ public static class Constants
 
     public static class Features
     {
+        public const string CustomData = "Custom Data";
+        public const string ComparatorSets = "Comparator Sets";
         public const string HealthCheck = "Health Check";
+        public const string FinancialPlans = "Financial Plans";
+        public const string UserData = "User Data";
     }
 }
