+++ platform/src/apis/Platform.Api.Benchmark/Features/ComparatorSets/DeleteSchoolUserDefinedComparatorSetFunction.cs
@@ -1,0 +1,58 @@
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Threading.Tasks;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
+using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
+using Microsoft.Extensions.Logging;
+using Platform.Api.Benchmark.Features.ComparatorSets.Services;
+using Platform.Functions.Extensions;
+using Platform.Functions.OpenApi;
+
+namespace Platform.Api.Benchmark.Features.ComparatorSets;
+
+public class DeleteSchoolUserDefinedComparatorSetFunction(IComparatorSetsService service, ILogger<DeleteSchoolUserDefinedComparatorSetFunction> logger)
+{
+    [Function(nameof(DeleteSchoolUserDefinedComparatorSetFunction))]
+    [OpenApiOperation(nameof(DeleteSchoolUserDefinedComparatorSetFunction), Constants.Features.ComparatorSets)]
+    [OpenApiParameter("urn", Type = typeof(string), Required = true)]
+    [OpenApiParameter("identifier", Type = typeof(string), Required = true)]
+    [OpenApiSecurityHeader]
+    [OpenApiResponseWithoutBody(HttpStatusCode.OK)]
+    [OpenApiResponseWithoutBody(HttpStatusCode.NotFound)]
+    [OpenApiResponseWithoutBody(HttpStatusCode.InternalServerError)]
+    public async Task<HttpResponseData> RunAsync(
+        [HttpTrigger(AuthorizationLevel.Admin, "delete", Route = Routes.SchoolUserDefinedComparatorSetItem)] HttpRequestData req,
+        string urn,
+        string identifier)
+    {
+        var correlationId = req.GetCorrelationId();
+
+        using (logger.BeginScope(new Dictionary<string, object>
+               {
+                   { "Application", Constants.ApplicationName },
+                   { "CorrelationID", correlationId },
+                   { "URN", urn },
+                   { "Identifier", identifier }
+               }))
+        {
+            try
+            {
+                var comparatorSet = await service.UserDefinedSchoolAsync(urn, identifier);
+                if (comparatorSet == null)
+                {
+                    return req.CreateNotFoundResponse();
+                }
+
+                await service.DeleteSchoolAsync(comparatorSet);
+                return req.CreateResponse(HttpStatusCode.OK);
+            }
+            catch (Exception e)
+            {
+                logger.LogError(e, "Failed to delete user defined school comparator set");
+                return req.CreateErrorResponse();
+            }
+        }
+    }
+}
