+++ platform/src/apis/Platform.Api.Benchmark/Features/ComparatorSets/GetSchoolDefaultComparatorSetFunction.cs
@@ -1,0 +1,52 @@
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Threading.Tasks;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
+using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
+using Microsoft.Extensions.Logging;
+using Platform.Api.Benchmark.Features.ComparatorSets.Services;
+using Platform.Api.Benchmark.OpenApi;
+using Platform.Functions.Extensions;
+using Platform.Functions.OpenApi;
+
+namespace Platform.Api.Benchmark.Features.ComparatorSets;
+
+public class GetSchoolDefaultComparatorSetFunction(IComparatorSetsService service, ILogger<GetSchoolDefaultComparatorSetFunction> logger)
+{
+    [Function(nameof(GetSchoolDefaultComparatorSetFunction))]
+    [OpenApiOperation(nameof(GetSchoolDefaultComparatorSetFunction), Constants.Features.ComparatorSets)]
+    [OpenApiParameter("urn", Type = typeof(string), Required = true)]
+    [OpenApiSecurityHeader]
+    [OpenApiResponseWithBody(HttpStatusCode.OK, "application/json", typeof(IComparatorSetSchool))]
+    [OpenApiResponseWithoutBody(HttpStatusCode.NotFound)]
+    [OpenApiResponseWithoutBody(HttpStatusCode.InternalServerError)]
+    public async Task<HttpResponseData> RunAsync(
+        [HttpTrigger(AuthorizationLevel.Admin, "get", Route = Routes.SchoolDefaultComparatorSet)] HttpRequestData req,
+        string urn)
+    {
+        var correlationId = req.GetCorrelationId();
+
+        using (logger.BeginScope(new Dictionary<string, object>
+               {
+                   { "Application", Constants.ApplicationName },
+                   { "CorrelationID", correlationId },
+                   { "URN", urn }
+               }))
+        {
+            try
+            {
+                var comparatorSet = await service.DefaultSchoolAsync(urn);
+                return comparatorSet == null
+                    ? req.CreateNotFoundResponse()
+                    : await req.CreateJsonResponseAsync(comparatorSet);
+            }
+            catch (Exception e)
+            {
+                logger.LogError(e, "Failed to get default school comparator set");
+                return req.CreateErrorResponse();
+            }
+        }
+    }
+}
