+++ platform/tests/Platform.Benchmark.Tests/FinancialPlans/WhenGetFinancialPlansRuns.cs
@@ -1,0 +1,47 @@
+using System.Net;
+using Microsoft.Extensions.Logging.Abstractions;
+using Moq;
+using Platform.Api.Benchmark.Features.FinancialPlans;
+using Platform.Api.Benchmark.Features.FinancialPlans.Services;
+using Platform.Test;
+using Xunit;
+
+namespace Platform.Benchmark.Tests.FinancialPlans;
+
+public class WhenGetFinancialPlansRuns : FunctionsTestBase
+{
+    private readonly GetFinancialPlansFunction _function;
+    private readonly Mock<IFinancialPlansService> _service;
+
+    public WhenGetFinancialPlansRuns()
+    {
+        _service = new Mock<IFinancialPlansService>();
+        _function = new GetFinancialPlansFunction(new NullLogger<GetFinancialPlansFunction>(), _service.Object);
+    }
+
+    [Fact]
+    public async Task ShouldReturn200OnValidRequest()
+    {
+        _service
+            .Setup(d => d.QueryAsync(It.IsAny<string[]>()))
+            .ReturnsAsync([]);
+
+        var result = await _function.RunAsync(CreateHttpRequestData());
+
+        Assert.NotNull(result);
+        Assert.Equal(HttpStatusCode.OK, result.StatusCode);
+    }
+
+    [Fact]
+    public async Task ShouldReturn500OnError()
+    {
+        _service
+            .Setup(d => d.QueryAsync(It.IsAny<string[]>()))
+            .Throws(new Exception());
+
+        var result = await _function.RunAsync(CreateHttpRequestData());
+
+        Assert.NotNull(result);
+        Assert.Equal(HttpStatusCode.InternalServerError, result.StatusCode);
+    }
+}
