+++ platform/tests/Platform.Benchmark.Tests/UserData/WhenGetUserDataFunctionRuns.cs
@@ -1,17 +1,33 @@
 using System.Net;
 using AutoFixture;
+using FluentValidation;
 using FluentValidation.Results;
 using Microsoft.Extensions.Primitives;
 using Moq;
+using Platform.Api.Benchmark.Features.UserData;
 using Platform.Api.Benchmark.Features.UserData.Parameters;
+using Platform.Api.Benchmark.Features.UserData.Services;
 using Platform.Functions;
+using Platform.Test;
 using Platform.Test.Extensions;
 using Xunit;
 
 namespace Platform.Benchmark.Tests.UserData;
 
-public class WhenFunctionReceivesQueryUserDataRequest : UserDataFunctionsTestBase
+public class WhenGetUserDataFunctionRuns : FunctionsTestBase
 {
+    private readonly Fixture _fixture = new();
+    private readonly GetUserDataFunction _function;
+    private readonly Mock<IUserDataService> _service;
+    private readonly Mock<IValidator<UserDataParameters>> _validator;
+
+    public WhenGetUserDataFunctionRuns()
+    {
+        _service = new Mock<IUserDataService>();
+        _validator = new Mock<IValidator<UserDataParameters>>();
+        _function = new GetUserDataFunction(_service.Object, _validator.Object);
+    }
+
     [Theory]
     [InlineData("userId", null, null, null, null, null)]
     public async Task ShouldReturn200OnValidRequest(string userId, string? type, string? status, string? id, string? organisationId, string? organisationType)
@@ -26,18 +42,18 @@ public class WhenFunctionReceivesQueryUserDataRequest : UserDataFunctionsTestBas
             { "organisationType", organisationType }
         };
 
-        var userData = Fixture.CreateMany<Api.Benchmark.Features.UserData.Models.UserData>();
+        var userData = _fixture.CreateMany<Api.Benchmark.Features.UserData.Models.UserData>();
 
-        Service
+        _service
             .Setup(d => d.QueryAsync(userId, type, status, id, organisationId, organisationType))
             .ReturnsAsync(userData)
             .Verifiable(Times.Once);
 
-        Validator
+        _validator
             .Setup(v => v.ValidateAsync(It.IsAny<UserDataParameters>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(new ValidationResult());
 
-        var result = await Functions.QueryAsync(CreateHttpRequestData(query));
+        var result = await _function.RunAsync(CreateHttpRequestData(query));
 
         Assert.NotNull(result);
         Assert.Equal(HttpStatusCode.OK, result.StatusCode);
@@ -46,20 +62,20 @@ public class WhenFunctionReceivesQueryUserDataRequest : UserDataFunctionsTestBas
         var actual = await result.ReadAsJsonAsync<Api.Benchmark.Features.UserData.Models.UserData[]>();
         Assert.NotNull(actual);
         Assert.Equivalent(userData, actual);
-        Service.Verify();
+        _service.Verify();
     }
 
 
     [Fact]
     public async Task ShouldReturn400OnValidationError()
     {
-        Validator
+        _validator
             .Setup(v => v.ValidateAsync(It.IsAny<UserDataParameters>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(new ValidationResult([
                 new ValidationFailure(nameof(UserDataParameters.UserId), "error message")
             ]));
 
-        var result = await Functions.QueryAsync(CreateHttpRequestData());
+        var result = await _function.RunAsync(CreateHttpRequestData());
 
         Assert.NotNull(result);
         Assert.Equal(HttpStatusCode.BadRequest, result.StatusCode);
@@ -69,7 +85,7 @@ public class WhenFunctionReceivesQueryUserDataRequest : UserDataFunctionsTestBas
         Assert.NotNull(values);
         Assert.Contains(values, p => p.PropertyName == nameof(UserDataParameters.UserId));
 
-        Service
+        _service
             .Setup(d => d.QueryAsync(It.IsAny<string>(), It.IsAny<string?>(), It.IsAny<string?>(), It.IsAny<string?>(), It.IsAny<string?>(), It.IsAny<string?>()))
             .Verifiable(Times.Never);
     }
